first working workflow
this is a bit of hack, but we deploy a local registry alongside
the cluster, and then use it to push / pull all previous dependent
job artifacts before running a step. This is a proof of concept
only because ideally we can replace this with a better temporary
storage strategy.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>