first successful run with Job! (#3)
* first successful run with Job!

There are a few underlying bugs in snakemake that need to be fixed,
but this runs fully for the first time! So exciting! Since the remote storage
with s3 is now working, I'm also removing the oras operator.
I think it adds unecessary complexity, despite the fact I like the 
project.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>