fix: hello world example
The example for hello-world was not working, I was trying to use
a hack for storage, and more appropriately I should use snakemake
built in remotes to start. This will run the entire hello world
example workflow, using a container I built that will eventually
be updated to snakemake releases when development is done.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>